1.拆分store与artfifactory相关业务接口，以api-artifactory-store插件方式接入到artifactory
