+++ src/backend/artifactory/api-artifactory-sample/src/main/kotlin/com/tencent/devops/artifactory/api/BuildFileResource.kt
