+++ src/backend/artifactory/api-artifactory-store/src/main/kotlin/com/tencent/devops/artifactory/pojo/ReArchiveAtomRequest.kt
