+++ src/backend/artifactory/biz-artifactory-sample/build.gradle
@@ -25,8 +25,6 @@
  */
 
 dependencies {
-    compile project(":common:common-auth:common-auth-sample")
-    compile project(":store:api-store-sample")
     compile project(":artifactory:api-artifactory-sample")
     compile project(":artifactory:biz-artifactory")
