+++ src/backend/artifactory/biz-artifactory-sample/src/main/kotlin/com/tencent/devops/artifactory/resources/BuildFileResourceImpl.kt
