+++ src/backend/artifactory/biz-artifactory-store/src/main/kotlin/com/tencent/devops/artifactory/resources/UserArchiveAtomResourceImpl.kt
