+++ src/backend/artifactory/boot-artifactory/build.gradle
@@ -25,8 +25,10 @@
  */
 
 dependencies {
+    compile project(":common:common-auth:common-auth-sample")
     compile project(":artifactory:biz-artifactory")
-    compile project(":artifactory:biz-artifactory-sample")
+    compile project(":artifactory:biz-artifactory-store") // 对接研发商店
+    compile project(":artifactory:biz-artifactory-sample") // sample 是开源版实现
 }
 
 apply from: "$rootDir/task_spring_boot_package.gradle"
