+++ src/backend/settings.gradle
@@ -72,8 +72,10 @@ include ":environment:model-environment"
 include ":artifactory"
 include ":artifactory:api-artifactory"
 include ":artifactory:api-artifactory-sample"
+include ":artifactory:api-artifactory-store"
 include ":artifactory:biz-artifactory"
 include ":artifactory:biz-artifactory-sample"
+include ":artifactory:biz-artifactory-store"
 include ":artifactory:boot-artifactory"
 include ":artifactory:model-artifactory"
