+++ src/backend/worker/worker-api-sdk/build.gradle
@@ -26,6 +26,7 @@
 
 dependencies {
     compile project(":worker:worker-common")
+    compile project(":artifactory:api-artifactory-store")
 }
 
 apply from: "$rootDir/task_deploy_to_maven.gradle"
